requirements.txt
numpy>=2.0.2
pandas>=2.2.3
scikit-learn>=1.6.1
matplotlib>=3.10.0
seaborn>=0.13.2
nltk>=3.9.1
pymorphy2>=0.9.1
pymorphy2-dicts-ru>=2.4.417127.4579844
keras>=3.8.0
pyyaml>=6.0
python-dotenv>=1.0
tqdm>=4.67.1
joblib>=1.4

# dev
# black>=24.8
# isort>=5.13
# flake8>=7.1
# mypy>=1.11
# pytest>=8.3
# pytest-cov>=5.0
# pre-commit>=3.8